step 1  in optimization, converting to native types of float instead of Double etc.

+++ src/main/java/com/searchbox/commons/params/SenseParams.java
@@ -11,7 +11,7 @@ package com.searchbox.commons.params;
 public class SenseParams {
     
     public static String SENSE_WEIGHT = "sw";
-    public static double DEFAULT_SENSE_WEIGHT = 1.0;
+    public static float DEFAULT_SENSE_WEIGHT = 1.0f;
 
     public static String SENSE_FIELD = "sf";
     public static String DEFAULT_SENSE_FIELD = "article_abstract";
