+++ src/main/java/com/searchbox/math/DoubleFullVector.java
@@ -9,9 +9,9 @@ package com.searchbox.math;
  * @author gamars
  */
 public class DoubleFullVector {
-       private double[] vector;
+       private float[] vector;
 
-    public DoubleFullVector(double[] vector) {
+    public DoubleFullVector(float[] vector) {
         this.vector = vector;
     }
 
@@ -19,34 +19,34 @@ public class DoubleFullVector {
         return vector.length == 0;
     }
 
-    public double getNorm() {
+    public float getNorm() {
         return VectorUtils.calculateNorm(vector);
     }
 
-    public double getDistance(final DoubleFullVector vector) {
+    public float getDistance(final DoubleFullVector vector) {
         return getDistance(vector.getData());
     }
 
-    public double getDistance(double vright[]) {
-        double result = 0;
-        double half;
+    public float getDistance(float vright[]) {
+        float result = 0;
+        float half;
         for (int zz = 0; zz < vector.length; zz++) {
             half = vector[zz] - vright[zz];
             result += half * half;
         }
-        return Math.sqrt(result);
+        return (float)Math.sqrt(result);
     }
 
     public int getDimension() {
         return vector.length;
     }
 
-    public double[] getData() {
+    public float[] getData() {
         return vector;
     }
 
-    public double dotProduct(final double[] weights) {
-        double result = 0;
+    public float dotProduct(final float[] weights) {
+        float result = 0;
         for (int zz = 0; zz < vector.length; zz++) {
             result += vector[zz] * weights[zz];
         }
@@ -54,7 +54,7 @@ public class DoubleFullVector {
     }
 
     public DoubleFullVector getSubVector(int start, int end) {
-        double out[] = new double[end - start];
+        float out[] = new float[end - start];
         for (int zz = start; zz < end; zz++) {
             out[zz - start] = vector[zz];
         }
@@ -62,10 +62,10 @@ public class DoubleFullVector {
     }
 
     public DoubleFullVector getUnitVector() {
-        double norm = getNorm();
-        double[] out = new double[vector.length];
+        float norm = getNorm();
+        float[] out = new float[vector.length];
         for (int zz = 0; zz < vector.length; zz++) {
-            out[zz] = vector[zz] * 1.0 / norm;
+            out[zz] = vector[zz] * 1.0f / norm;
         }
         return new DoubleFullVector(out);
     }
