+++ src/main/java/com/searchbox/math/VectorUtils.java
@@ -10,24 +10,24 @@ package com.searchbox.math;
  */
 public class VectorUtils {
 
-//    public static double calculateNorm(final TermFreqVector v) {
+//    public static float calculateNorm(final TermFreqVector v) {
 //        return calculateNorm(v.getTermFrequencies());
 //    }
     
-    public static double calculateNorm(final double[] vector) {
-        double norm = 0;
+    public static float calculateNorm(final float[] vector) {
+        float norm = 0;
         for (int zz = 0; zz < vector.length; zz++) {
             norm += vector[zz] * vector[zz];
         }
-        return Math.sqrt(norm);
+        return (float)Math.sqrt(norm);
     }
     
-    public static double calculateNorm(final int[] v) {
-        double norm = 0;
+    public static float calculateNorm(final int[] v) {
+        float norm = 0;
         for (int each : v) {
             norm += each * each;
         }
-        return Math.sqrt(norm);
+        return (float)Math.sqrt(norm);
     }
     
 //    public static long calculateSum(final TermFreqVector vector) {
